PatternLength 3, General gun rule, DeltaScan
The Pattern will be collected if at least 3 scans in row are detected. The new gun rule prevent the gun from working if the target was long time not seen. 

The deltaScan is now considered to shot at targets blind. If the scan is on something other the main target can still be shot at