new gun class - pattern and displacement vectors
This gun is a prof of concept for the new pattern state. The target collects pattern keys for his movement. 
The key is +0- Movement +0- rotate +0- acceleration. The key represents a overall movement profile and can adapt very quick. There are no values included, just the state. This makes the key very small an easy to generate. It needs more research to use the key for fire angle calculation. My thought was to implement some sort of displacement vectors saved along with the keys and just take the vectors to generate a fire angle. 
First tests show the key looks very sound but random movement is still something have to be concerned. 
I played a little with searching trough the pattern back and forth to find the most used pattern within the current view.