version 4.6 changes
This is mainly a switchback to v4.4 but with some neat improvements to the movement (hopefully). The movement distance for the min risk is now based on the distance and the 'predicted' bullet power. Basically this means i don't travel as much as he can hit me if he uses linear targeting.