Tassie v1.8.? messed up code
Tried some wired ideas with the movement and messed up the code completely. I have to go back for more seriously tests. I just wanted to see if something could just by accident bring something good. The basic idea was to bring the minion in a better combat position to guard the leader just by being closer to the enemy. The movement got some distance related adjustments and it might be rewarding to recheck this. 

I'm convinced i should invest in a better gun, even if this means less tactic.